requirements.txt
streamlit>=1.33
spacy>=3.7
pdfminer.six>=20231228
docx2txt>=0.8
